Initial Commit
Building a bubble shooter game. Will be themed around jollof rice
(Google it if you don’t understand). Initial commit lays out the game
area, the header and the modal for a new game. The is just a sketch.
Will be focusing on the functionality first before going in depth on
design.